moved tabbed associations to the events (task #3007)
