+++ src/Template/Element/associated_records.ctp
@@ -4,7 +4,6 @@ use Cake\Utility\Inflector;
 use CsvMigrations\FieldHandlers\FieldHandlerFactory;
 
 $fhf = new FieldHandlerFactory();
-
 $panels = [];
 if (!empty($csvAssociatedRecords['oneToMany'])) {
     foreach ($csvAssociatedRecords['oneToMany'] as $tabName => $assocData) {
@@ -15,7 +14,6 @@ if (!empty($csvAssociatedRecords['oneToMany'])) {
         }
     }
 }
-
 /*
 list of embedded fields to generate modals from
  */
@@ -30,7 +28,6 @@ if (!empty($csvAssociatedRecords['manyToMany'])) {
     }
 }
 ?>
-
 <?php if (!empty($panels)) : ?>
 <div class="row associated_records">
     <div class="col-xs-12">
@@ -43,13 +40,18 @@ if (!empty($csvAssociatedRecords['manyToMany'])) {
             <li role="presentation" class="<?= $active; ?>">
                 <a href="#<?= $tabName; ?>" aria-controls="<?= $tabName; ?>" role="tab" data-toggle="tab">
                     <?php
-                        $tableName = Inflector::humanize($assocData['table_name']);
-                        $fieldName = trim(str_replace($tableName, '', Inflector::humanize(Inflector::tableize($tabName))));
-                        if (!empty($fieldName)) {
-                            $fieldName = ' <small>(' . $fieldName . ')</small>';
+                        //prettifying the tabs names
+                        if (!empty($csvAssociationLabels) && in_array($tabName, array_keys($csvAssociationLabels))) {
+                            echo $csvAssociationLabels[$tabName];
+                        } else {
+                            $tableName = Inflector::humanize($assocData['table_name']);
+                            $fieldName = trim(str_replace($tableName, '', Inflector::humanize(Inflector::tableize($tabName))));
+                            if (!empty($fieldName)) {
+                                $fieldName = '<small>(' . $fieldName . ')</small>';
+                            }
+                            echo $tableName . $fieldName;
                         }
                     ?>
-                    <?= $tableName . $fieldName ?>
                 </a>
             </li>
 <?php
@@ -61,7 +63,7 @@ if (!empty($csvAssociatedRecords['manyToMany'])) {
 <?php
     $active = 'active';
     foreach ($panels as $assocName => $assocData) {
-    ?>
+?>
             <div role="tabpanel" class="tab-pane <?= $active; ?>" id="<?= $assocName; ?>">
             <?php
             /*
